symbol-extract/requirements.txt
google-cloud-storage==1.26.0
retry==0.9.2